Move CPossessionState traits
