+++ include/football/CPossessionState.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "IJsonable.h"
-#include "traits/json.h"
+#include "football/traits/CPossessionState.h"
 
 #include "football/types/CPossessionState.h"
 
@@ -13,27 +13,7 @@
 
 #include "ExceptionUtils.h"
 
-namespace futsim
-{
-
-namespace football
-{
-class CPossessionState;
-}
-
-template <> struct json_traits<football::CPossessionState>
-{
-	//! JSON key for the class.
-	static inline constexpr std::string_view KEY = "Possession state";
-	//! JSON key for the \copybrief football::CPossessionState::mOutcome
-	static inline constexpr std::string_view OUTCOME_KEY = "Outcome";
-	//! JSON key for the \copybrief football::CPossessionState::mTackler
-	static inline constexpr std::string_view TACKLER_KEY = "Tackler";
-	//! JSON key for the \copybrief football::CPossessionState::mPasser
-	static inline constexpr std::string_view PASSER_KEY = "Passer";
-};
-
-namespace football
+namespace futsim::football
 {
 
 /**
@@ -139,6 +119,4 @@ CPossessionState::CPossessionState(
 }
 FUTSIM_CATCH_AND_RETHROW_EXCEPTION( std::invalid_argument, "Error creating the possession state." )
 
-} // football namespace
-
-} // futsim namespace
+} // futsim::football namespace
